+++ src/main/java/com/aquaticinformatics/aquarius/sdk/samples/serializers/OffsetDateTimeDeserializer.java
@@ -6,29 +6,32 @@ import com.google.gson.JsonElement;
 import com.google.gson.JsonParseException;
 
 import java.lang.reflect.Type;
-import java.time.Instant;
+import java.time.OffsetDateTime;
 import java.time.format.DateTimeFormatter;
 import java.time.format.DateTimeParseException;
 import java.util.Locale;
 
-public class InstantDeserializer implements JsonDeserializer<Instant> {
 
-    public static final String ZoneFieldPattern = "ZZZZZ";
-    public static final String Pattern = "yyyy'-'MM'-'dd'T'HH':'mm':'ss'.'SSS" + ZoneFieldPattern;
+public class OffsetDateTimeDeserializer implements JsonDeserializer<OffsetDateTime> {
+
+    public static final String ZoneFieldPattern = "XXXXX";
+    public static final String DateAndTimePattern = "yyyy'-'MM'-'dd'T'HH':'mm':'ss";
+    public static final String FractionalSecondsPattern = "'.'SSS";
+    public static final String Pattern = DateAndTimePattern + FractionalSecondsPattern + ZoneFieldPattern;
 
     public static final DateTimeFormatter FORMATTER = DateTimeFormatter
             .ofPattern(Pattern)
             .withLocale(Locale.US);
-    public static final String JsonMaxValue = "MaxInstant";
-    public static final String JsonMinValue = "MinInstant";
-    public static final Instant MaxValue = Instant.MAX;
-    public static final Instant MinValue = Instant.MIN;
-    public static final String JsonMaxConcreteValue = "9999-12-31T23:59:59.999Z";
-    public static final String JsonMinConcreteValue = "0001-01-01T00:00:00.000Z";
-    public static final Instant MaxConcreteValue = FORMATTER.parse(JsonMaxConcreteValue, Instant::from);
-    public static final Instant MinConcreteValue = FORMATTER.parse(JsonMinConcreteValue, Instant::from);
+    public static final String JsonMaxValue = "MaxOffsetDateTime";
+    public static final String JsonMinValue = "MinOffsetDateTime";
+    public static final OffsetDateTime MaxValue = OffsetDateTime.MAX;
+    public static final OffsetDateTime MinValue = OffsetDateTime.MIN;
+    public static final String JsonMaxConcreteValue = "9999-12-31T23:59:59.999-18:00";
+    public static final OffsetDateTime MaxConcreteValue = FORMATTER.parse(JsonMaxConcreteValue, OffsetDateTime::from);
+    public static final String JsonMinConcreteValue = "0001-01-01T00:00:00.000+18:00";
+    public static final OffsetDateTime MinConcreteValue = FORMATTER.parse(JsonMinConcreteValue, OffsetDateTime::from);
 
-    public static Instant parse(String text) {
+    public static OffsetDateTime parse(String text) {
         if (text.equalsIgnoreCase(JsonMaxValue))
             return MaxValue;
 
@@ -46,14 +49,14 @@ public class InstantDeserializer implements JsonDeserializer<Instant> {
             // Workaround for quirks: 2020-12-01T00:00:00.000
             if (!text.matches(".*([+-]\\d{2}:?\\d{2}|Z)$"))
                 text = text + "Z";
-            return FORMATTER.parse(text, Instant::from);
+            return FORMATTER.parse(text, OffsetDateTime::from);
         } catch (DateTimeParseException exception) {
             throw (exception);
         }
     }
 
     @Override
-    public Instant deserialize(JsonElement jsonElement, Type type, JsonDeserializationContext context) throws JsonParseException {
+    public OffsetDateTime deserialize(JsonElement jsonElement, Type type, JsonDeserializationContext context) throws JsonParseException {
         return parse(jsonElement.getAsString());
     }
 }