+++ src/test/java/com/aquaticinformatics/aquarius/sdk/samples/serializers/InstantDeserializerTest.java
@@ -30,7 +30,7 @@ public class InstantDeserializerTest {
     public ExpectedException thrown = ExpectedException.none();
 
     @Test
-    public void parse_WithNullText_Throws(){
+    public void parse_WithNullText_Throws() {
         thrown.expect(NullPointerException.class);
 
         InstantDeserializer.parse(null);
@@ -51,12 +51,15 @@ public class InstantDeserializerTest {
                 new Object[]{"Xmas 2017 PST with timezone datum code", "2017-12-25T00:00:00.123-07:00[PST]", Instant.parse("2017-12-25T07:00:00.123Z")},
                 new Object[]{"Missing time component assumes midnight (Vancouver)", "2020-12-01T-07:00", Instant.parse("2020-12-01T07:00:00.000Z")},
                 new Object[]{"Missing time component assumes midnight (Sydney)", "2020-12-01T+10:00", Instant.parse("2020-11-30T14:00:00.000Z")},
+                new Object[]{"Missing offset component assumes UTC", "2020-12-01T00:00:00.000", Instant.parse("2020-12-01T00:00:00.000Z")},
+                new Object[]{"Missing both offset and time component assumes midnight and UTC", "2000-07-15T", Instant.parse("2000-07-15T00:00:00.000Z")},
+                new Object[]{"Missing time component and offset is UTC assumes midnight", "2000-07-15TZ", Instant.parse("2000-07-15T00:00:00.000Z")},
         };
     }
 
     @Test
     @Parameters(method = "validInstantTests")
-    public void parse_WithValidInstant_ConvertsAsExpected(String reason, String instantText, Instant expected){
+    public void parse_WithValidInstant_ConvertsAsExpected(String reason, String instantText, Instant expected) {
 
         Instant actual = InstantDeserializer.parse(instantText);
 
@@ -65,7 +68,7 @@ public class InstantDeserializerTest {
 
     @Test
     @Parameters(method = "validInstantTests")
-    public void fromJson_WithValidInstant_ConvertsAsExpected(String reason, String instantText, Instant expected){
+    public void fromJson_WithValidInstant_ConvertsAsExpected(String reason, String instantText, Instant expected) {
 
         String json = "\"" + instantText + "\"";