+++ ReleaseNotes.md
@@ -4,6 +4,10 @@ This page highlights some changes in the SDK.
 
 Not all changes will be listed, but you can always [compare by version tags](https://github.com/AquaticInformatics/aquarius-sdk-java/compare/v17.2.26...v17.2.28) to see the full source code difference.
 
+### 25.3.0
+- Fixed vulnerability by upgrading Apache Tika.
+- Updated service models for the AQUARIUS Samples 2025.10 release.
+
 ### 25.2.0
 - Updated service models for the AQUARIUS Time-Series 2025.2 release.